backend/requirements.txt
Celery
Django
djangorestframework
psycopg2